refactor!: move nip19 to own module nostr-nip19
