+++ examples/nostr-client-example-application/build.gradle
@@ -5,6 +5,7 @@ description = 'nostr client example application package'
 dependencies {
     implementation project(':nostr:nostr-core')
     implementation project(':nostr:nostr-client')
+    implementation project(':nostr:nostr-nip19')
 
     implementation "fr.acinq.bitcoin:bitcoin-kmp-jvm:${acinqBitcoinKmpVersion}"
     implementation "fr.acinq.secp256k1:secp256k1-kmp-jni-jvm:${acinqSecp256k1KmpVersion}"
