+++ examples/nostr-shell-example-application/build.gradle
@@ -14,6 +14,7 @@ bootRun {
 
 dependencies {
     implementation project(':nostr:nostr-core')
+    implementation project(':nostr:nostr-nip19')
 
     implementation "fr.acinq.bitcoin:bitcoin-kmp-jvm:${acinqBitcoinKmpVersion}"
     implementation "fr.acinq.secp256k1:secp256k1-kmp-jni-jvm:${acinqSecp256k1KmpVersion}"
