+++ nostr-relay/nostr-relay-starter/build.gradle
@@ -16,6 +16,8 @@ dependencies {
 
     api project(':nostr-relay:nostr-relay-autoconfigure')
 
+    api project(':nostr:nostr-nip19')
+
     implementation 'org.springframework.boot:spring-boot-starter-websocket'
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.springframework.boot:spring-boot-starter-security'
