+++ settings.gradle
@@ -6,6 +6,7 @@ include 'nostr:nostr-client'
 include 'nostr:nostr-identity'
 include 'nostr:nostr-nip6'
 include 'nostr:nostr-nip11'
+include 'nostr:nostr-nip19'
 include 'nostr:nostr-proto'
 include 'nostr:nostr-proto-json'
 include 'nostr:nostr-template'
